Commits template filestructure for puppet using hiera and librian-puppet
This allows PS:1 to use puppet to manage servers and services. Current
structure is bare. Modules are loaded by editing the Puppetfile and re-running
librarian-puppet. This allows management of modules directly from upstream.

Hiera is a hierachicial data lookup tool. This allows for the lookup
precedence to be specified in hiera.yaml. Using the hiera tool within
your manifests allows a machine to lookup common datastructures based on
said precedence. Hiera more closely matches the structure that will be
coming in Puppet 3.0. Read more here:

http://projects.puppetlabs.com/projects/hiera